requirements.txt
fastapi
uvicorn
spacy
fastapi-sqlalchemy
sqlalchemy
langchain
langchain-openai
python-dotenv
pydantic
pandas
openai
faiss-cpu
jinja2
